+++ R/NMscanData.R
@@ -149,8 +149,8 @@
 ##'
 ##' \describe{
 ##'  \item{character TIME}{If Nonmem is used to translate DAY and a character TIME column, TIME has to be available in an output table. NMscanData does not do the translation to numeric.}
-##'  \item{RECORDS}{The RECORDS option to limit the part of the input data being used is not searched for. Using merge.by.row=TRUE will work unaffectedly.}
-##'  \item{NULL}{The NULL argument to specify missing value string in input data is not respected. If delimited input data is read (as opposed to rds files), missing values are assumed to be represented by dots (.).}
+##'  \item{RECORDS}{The RECORDS option to limit the part of the input data being used is not searched for. Using `merge.by.row=TRUE` will work unaffectedly.}
+##'  \item{NULL}{The NULL argument to specify missing value string in input data is not respected. If delimited input data is read (as opposed to rds files), missing values are assumed to be represented by dots (.). This is used for conversion of missing values to `NA` before returning to user.}
 ##' }
 ##' @examples
 ##' \dontrun{