+++ R/dtFillObs.R
@@ -1,5 +1,10 @@
-##' Create a data table with columns of unequal length
+##' Create a data table with input columns of unequal length
 ##'
+##' All columns are filled with NA to max length of all input variables. This is
+##' useful when constructing partially completed variables where 1) variables
+##' are complete up to their length and 2) the missing values should be set to
+##' NA.
+##' 
 ##' @param ... Vectors to put into data.table
 ##'
 ##' @keywords internal