+++ inst/examples/scripts/run_models.R
@@ -1,4 +1,0 @@
-library(NMsim)
-
-NMexec("../nonmem/xgxr132.mod",sge=FALSE)
-NMexec("../nonmem/xgxr134.mod",sge=FALSE)