+++ tests/testthat/test_NMapplyFilters.R
@@ -1,2 +1,0 @@
-## test filters on NA values
-