+++ tests/testthat/test_NMinfo.R
@@ -5,7 +5,7 @@ test_that("basic",{
     file.lst <- system.file("examples/nonmem/xgxr001.lst" ,package="NMdata")
     res1 <- NMscanData(file=file.lst,quiet=T,order.columns = F,merge.by.row=FALSE,check.time=F)
 
-    info.cols <- NMinfo(res1,info="columns")
+    info.cols <- NMinfo(res1,info="columns",as.fun=as.data.frame)
     
     expect_equal_to_reference(info.cols,fileRef,version=2)